board/main.py: load modules async while loop is already running
The purpose of loading modules async is to speed-up the time the board takes to become
operational. This is particularly useful for systems that have a display because the display
can show a splash screen rapidly at reset (if the appropriate module is placed first in the
list) as opposed to having to wait forever until the whole universe is loaded.